requirements.txt
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
python-dotenv==1.0.0
requests==2.31.0
django==4.2.4
django-modeltranslation==0.18.11
django-parler==2.3
# itypes
# Jinja2
# MarkupSafe
# pycparser
# pyparsing
# DateTime
# tzdata
# zope.interface
# django-utils-six
# cffi
# coreapi
# coreschema